Corrected errors checkStyleTest
